+++ pkgs/test_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.6.16-wip
+
+* Add `SuiteConfiguration.suiteLoadTimeout` to configure the timeout for loading a test suite.
+* Removed hard-coded timeout of 12m for loading a test suite and set default to `none`.
+
 ## 0.6.15
 
 * Add `--coverage-package` flag, which filters the coverage report to specific